+++ CHANGELOG.md
@@ -43,6 +43,8 @@ Breaking changes should be prefixed by `[**BREAKING**]` (without the quotes), to
 
 - Server:
   - Notes now support extra JSON-LD @context provided in `custom`.
+  - Nodeinfo software name format now conforms to spec.
+  - Fixed JSON serialization bug on MySQL/MariaDB.
 - Tasks:
   - Replace factory names in `federails:copy_factories`